requirements.txt
python-dotenv==0.10.1
psycopg2==2.7.7
pandas==0.24.1
matplotlib
